+++ internal/provider/vendor_google_vertex.go
@@ -1,0 +1,9 @@
+package provider
+
+func init() {
+	RegisterVendorAdapter(simpleVendorAdapter{
+		name:       "google-vertex",
+		domains:    []string{"aiplatform.googleapis.com"},
+		defaultAPI: "google-vertex",
+	})
+}